Move create scenario hooks to new directory
Renamed and relocated the create scenario form hooks from 'components/Project/CreateScenarioForms/hooks.js' to 'hooks/create-scenario-forms.js'. Updated import paths in CreateScenario.jsx to reflect the new location. This improves project structure and organization.
